requirements.txt
streamlit
numpy
pypdf
python-dotenv
sentence-transformers==2.2.2
ollama